Move synset list downloading into its own script
In upcoming commits, the scripts for downloading and processing ImageNet
based off of the URL lists on the image-net.org website will be moved into
their own directory. There will then be a separate directory for processing
ImageNet based off the single link download that image-net.org provides. The
synset list downloading will remain common, though.
